+++ lib/duckdb/result.rb
@@ -28,7 +28,6 @@ module DuckDB
 
     TO_METHODS = if Gem::Version.new(DuckDB::LIBRARY_VERSION) == Gem::Version.new('0.10.0')
                    Hash.new(:_to_string).merge(
-                     1 => :_to_boolean,
                      3 => :_to_smallint,
                      4 => :_to_integer,
                      5 => :_to_bigint,
@@ -40,7 +39,6 @@ module DuckDB
                    ).freeze
                  else
                    Hash.new(:_to_string).merge(
-                     1 => :_to_boolean,
                      3 => :_to_smallint,
                      4 => :_to_integer,
                      5 => :_to_bigint,